+++ ClassMonitor/plugins/power.lua
@@ -8,13 +8,10 @@ local PixelPerfect = Engine.PixelPerfect
 local DefaultBoolean = Engine.DefaultBoolean
 local GetColor = Engine.GetColor
 
--- self.count = max power ever
--- self.maxValue = current max power
-
---
 local plugin = Engine:NewPlugin("POWER")
---print(tostring(UnitPower("player", self.settings.powerType)))
+
 -- own methods
+
 function plugin:UpdateVisibility(event)
 	--
 --print("POWER:UpdateVisibility")
@@ -26,136 +23,71 @@ function plugin:UpdateVisibility(event)
 	end
 	--
 	if (self.settings.autohide == false or inCombat) and CheckSpec(self.settings.specs) then
---print("SHOW")
-		--
 		--self:UpdateMaxValue()
 		self:UpdateValue()
 		--
 		self.frame:Show()
 	else
---print("HIDE")
 		self.frame:Hide()
 	end
 end
 
---[[
-function plugin:UpdateMaxValue(event, unit, powerType)
-	--
---print("POWER:UpdateMaxValue:"..tostring(event).."  "..tostring(unit).."  "..tostring(powerType))
-	local maxValue = UnitPowerMax("player", self.settings.powerType)
---print("MAX:"..tostring(event).."  maxValue:"..tostring(maxValue).."  count:"..tostring(self.count).."  max:"..tostring(self.maxValue))
-	if maxValue ~= self.maxValue then
-		-- compute new width, spacing
-		--local width, spacing = PixelPerfect(self.settings.width, maxValue)
-		local width, spacing = PixelPerfect(self:GetWidth(), maxValue)
-		if maxValue > self.count then
-			self.count = maxValue
+function plugin:UpdateValue(event, unit, powerType)
+	local value = UnitPower("player", self.settings.powerType)
+	if self.settings.borderRemind == true then
+		if value and value > 0 then
+			for i = 1, value do self.points[i].status:Show() end
+			for i = value+1, self.count do self.points[i].status:Hide() end
+		else
+			for i = 1, self.count do self.points[i].status:Hide() end
 		end
-		for i = 1, self.count do
-			self:UpdatePointGraphics(i, width, spacing)
-			self.points[i]:Hide()
+	else
+		if value and value > 0 then
+	--		assert(value <= self.count, "Current value:"..tostring(value).." must be <= to count:"..tostring(self.count))
+			for i = 1, value do self.points[i]:Show() end
+			for i = value+1, self.count do self.points[i]:Hide() end
+		else
+			for i = 1, self.count do self.points[i]:Hide() end
 		end
-
--- --print("WIDTH:"..tostring(width).."  SPACING:"..tostring(spacing))
-		-- -- create new points if needed
-		-- if maxValue > self.count then
-			-- for i = self.count+1, maxValue do
-				-- self:CreatePoint(i, width, spacing)
-			-- end
-			-- self.count = maxValue
-		-- end
-		-- -- hide all points
-		-- for i = 1, self.count do
-			-- self.points[i]:Hide()
-		-- end
-		-- -- resize points
-		-- for i = 1, maxValue do
-			-- local point = self.points[i]
-			-- point:Size(width, self.settings.height)
-			-- if i ~= 1 then -- first point doesn't move
-				-- point:ClearAllPoints()
-				-- point:Point("LEFT", self.points[i-1], "RIGHT", spacing, 0)
-			-- end
-		-- end
-
-		self.maxValue = maxValue
-		--
-		self:UpdateValue()
 	end
 end
 
-function plugin:UpdateValue(event, unit, powerType)
-	--
-	--if powerType ~= "CHI" then return end
---print("POWER:UpdateValue:"..tostring(event).."  "..tostring(unit).."  "..tostring(powerType))
-	local value = UnitPower("player", self.settings.powerType)
---print("CURRENT:"..tostring(value).."  count:"..tostring(self.count).."  max:"..tostring(self.maxValue))
-	if value and value > 0 then
-		--assert(value <= self.maxValue and value <= self.count, "Current value:"..tostring(value).." should be <= to maxValue:"..tostring(self.maxValue).." and <= to count:"..tostring(self.count))
-		assert(value <= self.count, "Current value:"..tostring(value).." must be <= to count:"..tostring(self.count))
-		for i = 1, value do self.points[i]:Show() end
-		for i = value+1, self.count do self.points[i]:Hide() end
-	else
-		for i = 1, self.count do self.points[i]:Hide() end
-	end
-end
---]]
-
-function plugin:UpdateValue(event, unit, powerType)
--- TODO: only for monitored power type     parameter powerType is a string (MANA, FOCUS, CHI, ...) but UnitPower/UnitPowerMax use an id
---[[
-	-- only for monitored power type
-	local resource = UnitPowerType("player", powerType)
-print("POWERTYPE:"..tostring(powerType).."  "..tostring(self.settings.powerType).."  "..tostring(resource))
-	if resource ~= self.settings.powerType then return end
---]]
-	--
+function plugin:SetCounts()
 	local maxValue = UnitPowerMax("player", self.settings.powerType)
-	if maxValue and maxValue > 0 and maxValue ~= self.maxValue then
-		-- compute new width, spacing
-		if maxValue > self.count then
-			self.count = maxValue
-		end
-		-- update points (create any needed points)
-		local width, spacing = PixelPerfect(self:GetWidth(), maxValue)
-		local height = self:GetHeight()
-		for i = 1, self.count do
-			self:UpdatePointGraphics(i, width, height, spacing)
-			self.points[i]:Hide()
-		end
-		-- update current max
-		self.maxValue = maxValue
-	end
-	--
-	local value = UnitPower("player", self.settings.powerType)
-	if value and value > 0 then
-		assert(value <= self.count, "Current value:"..tostring(value).." must be <= to count:"..tostring(self.count))
-		for i = 1, value do self.points[i]:Show() end
-		for i = value+1, self.count do self.points[i]:Hide() end
-	else
-		for i = 1, self.count do self.points[i]:Hide() end
+	print(maxValue)
+	if maxValue and maxValue ~= self.maxValue then
+		self.count = maxValue
+
+		self:UpdateGraphics()
 	end
 end
 
 function plugin:UpdatePointGraphics(index, width, height, spacing)
-	--
 	local point = self.points[index]
 	if not point then
 		point = CreateFrame("Frame", nil, self.frame)
 		point:SetTemplate()
 		point:SetFrameStrata("BACKGROUND")
-		point:Hide()
+--		point:Hide()
+		if not self.settings.borderRemind == true then point:Hide() end
 		self.points[index] = point
 	end
 	point:Size(width, height)
 	point:ClearAllPoints()
-	if index == 1 then
-		point:Point("TOPLEFT", self.frame, "TOPLEFT", 0, 0)
+	if self.settings.reverse == true then
+		if index == 1 then
+			point:Point("TOPRIGHT", self.frame, "TOPRIGHT", 0, 0)
+		else
+			point:Point("RIGHT", self.points[index-1], "LEFT", -spacing, 0)
+		end
 	else
-		point:Point("LEFT", self.points[index-1], "RIGHT", spacing, 0)
+		if index == 1 then
+			point:Point("TOPLEFT", self.frame, "TOPLEFT", 0, 0)
+		else
+			point:Point("LEFT", self.points[index-1], "RIGHT", spacing, 0)
+		end
 	end
-	--
-	if self.settings.filled == true and not point.status then
+	if not point.status then
 		point.status = CreateFrame("StatusBar", nil, point)
 		point.status:SetStatusBarTexture(UI.NormTex)
 		point.status:SetFrameLevel(6)
@@ -165,11 +97,13 @@ function plugin:UpdatePointGraphics(index, width, height, spacing)
 	local color = GetColor(self.settings.colors, index, UI.ClassColor())
 	if self.settings.filled == true then
 		point.status:SetStatusBarColor(unpack(color))
-		point.status:Show()
 		point:SetBackdropBorderColor(unpack(UI.BorderColor))
+		point.status:Show()
 	else
+		point.status:SetStatusBarColor(0, 0, 0, 1)
 		point:SetBackdropBorderColor(unpack(color))
-		if point.status then point.status:Hide() end
+		point.status:Show()
+--		if point.status then point.status:Hide() end
 	end
 end
 
@@ -187,7 +121,7 @@ function plugin:UpdateGraphics()
 	frame:Point(unpack(self:GetAnchor()))
 	frame:Size(frameWidth, height)
 	-- Create points
-	local width, spacing = PixelPerfect(frameWidth, self.maxValue)
+	local width, spacing = PixelPerfect(frameWidth, self.count)
 	self.points = self.points or {}
 	for i = 1, self.count do
 		self:UpdatePointGraphics(i, width, height, spacing)
@@ -197,15 +131,11 @@ end
 -- overridden methods
 function plugin:Initialize()
 	-- set defaults
-	self.settings.count = self.settings.count or 1 -- starts with count = 1 if count not found in settings
 	self.settings.filled = DefaultBoolean(self.settings.filled, false)
 	self.settings.powerType = self.settings.powerType or SPELL_POWER_HOLY_POWER --
 	self.settings.colors = self.settings.colors or self.settings.color or UI.PowerColor(self.settings.powerType) or UI.ClassColor()
-	--
-	self.count = self.settings.count
-	self.maxValue = self.count -- current max value (<= count)
-	-- Create a frame including every points
-	self:UpdateGraphics()
+
+	self:SetCounts()
 end
 
 function plugin:Enable()
@@ -215,8 +145,7 @@ function plugin:Enable()
 	self:RegisterUnitEvent("PLAYER_SPECIALIZATION_CHANGED", "player", plugin.UpdateVisibility)
 
 	self:RegisterUnitEvent("UNIT_POWER", "player", plugin.UpdateValue)
-	--self:RegisterUnitEvent("UNIT_MAXPOWER", "player", plugin.UpdateMaxValue)   if a monk is at 4 chi and learn talent Ascension  UNIT_POWER (5 chi) will be called before UNIT_MAXPOWER (max 5 chi)
-	self:RegisterUnitEvent("UNIT_MAXPOWER", "player", plugin.UpdateValue)
+	self:RegisterUnitEvent("UNIT_MAXPOWER", "player", plugin.SetCounts)
 end
 
 function plugin:Disable()